feat: add step1 speech lm and streamer
