+++ src/core/llm/transformers/manual_speech_llasa.py
@@ -1,12 +1,9 @@
 import logging
 from threading import Thread
-from queue import Queue
-
 
 try:
     import torch
     from transformers import AutoTokenizer, AutoModelForCausalLM
-    from transformers.generation.streamers import BaseStreamer
 except ModuleNotFoundError as e:
     logging.error(f"Exception: {e}")
     logging.error(
@@ -14,11 +11,11 @@ except ModuleNotFoundError as e:
     )
     raise Exception(f"Missing module: {e}")
 
-
 from src.common.utils.helper import get_device
 from src.common.session import Session
 from src.types.llm.transformers import TransformersSpeechLMArgs
 from .base import TransformersBaseLLM
+from .streamer import TokenStreamer
 
 
 def ids_to_speech_tokens(speech_ids):
@@ -39,43 +36,6 @@ def extract_speech_ids(speech_tokens_str):
         else:
             logging.error(f"Unexpected token: {token_str}")
     return speech_ids
-
-
-class TokenStreamer(BaseStreamer):
-    def __init__(self, skip_prompt: bool = False, timeout=None):
-        self.skip_prompt = skip_prompt
-
-        # variables used in the streaming process
-        self.token_queue = Queue()
-        self.stop_signal = None
-        self.next_tokens_are_prompt = True
-        self.timeout = timeout
-
-    def put(self, value):
-        if len(value.shape) > 1 and value.shape[0] > 1:
-            raise ValueError("TextStreamer only supports batch size 1")
-        elif len(value.shape) > 1:
-            value = value[0]
-
-        if self.skip_prompt and self.next_tokens_are_prompt:
-            self.next_tokens_are_prompt = False
-            return
-
-        for token in value.tolist():
-            self.token_queue.put(token)
-
-    def end(self):
-        self.token_queue.put(self.stop_signal)
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        value = self.token_queue.get(timeout=self.timeout)
-        if value == self.stop_signal:
-            raise StopIteration()
-        else:
-            return value
 
 
 class TransformersManualSpeechLlasa(TransformersBaseLLM):
